refactor: split into more components, also adds styling to stars in header
