+++ src/views/VorfreudenView.vue
@@ -1,9 +1,11 @@
 <template>
   <main>
+    <h2>Meine Vorfreuden heute</h2>
     <TheVorfreuden
       singleNote="single-anticipation"
       placeholderText="Hier die Vorfreude eingeben und mit Enter speichern."
     />
+    <AddCalendar placeholderText="Hier die Vorfreude eingeben und mit Enter speichern." />
     <InfoBlurb
       adjusting="adjusting-anticipation"
       text="Info: Speicher deine fertig geschriebene Vorfreude mit Enter"
@@ -15,11 +17,12 @@
 </template>
 
 <script>
-import TheVorfreuden from '../components/TheVorfreuden.vue'
+import TheVorfreuden from '../components/VorfreudenViewComponents/TheVorfreuden.vue'
 import InfoBlurb from '../components/InfoBlurb.vue'
+import AddCalendar from '../components/VorfreudenViewComponents/AddCalendar.vue'
 
 export default {
-  components: { TheVorfreuden, InfoBlurb },
+  components: { TheVorfreuden, InfoBlurb, AddCalendar },
   props: ['singleNote', 'placeholderText', 'text', 'adjusting']
 }
 </script>
@@ -33,6 +36,18 @@ main {
   background-color: #020b30;
   border-radius: 2rem;
   overflow: hidden;
+  padding: 2rem;
+}
+
+h2 {
+  color: #ec635e;
+  font-family: 'Antonio';
+  /* background-color: rgb(236, 99, 94, 0.1); */
+  /* background-color: rgb(79, 101, 223, 0.1); */
+  padding-bottom: 1rem;
+  /* border-radius: 4px; */
+  text-align: center;
+  margin-bottom: 1rem;
 }
 
 .btn-wrapper {
